+++ scarlet/source.py
@@ -3,7 +3,6 @@ import numpy as np
 from .component import Component, ComponentTree
 from . import operator
 from . import update
-from . import observation
 
 import logging
 logger = logging.getLogger("scarlet.source")
@@ -73,6 +72,40 @@ def get_best_fit_seds(morphs, scene, observations):
         seds[:, band:band+obs.B] = sed
     return seds
 
+def build_detection_coadd(sed, bg_rms, observation, scene, thresh=1):
+    """Build a band weighted coadd to use for source detection
+
+    Parameters
+    ----------
+    sed: array
+        SED at the center of the source.
+    bg_rms: array
+        Background RMS in each band in observation.
+    observation: `~scarlet.observation.Observation`
+        Observation to use for the coadd.
+    scene: `scarlet.observation.Scene`
+        The scene that the model lives in.
+    thresh: `float`
+        Multiple of the backround RMS used as a
+        flux cutoff.
+
+    Returns
+    -------
+    detect: array
+        2D image created by weighting all of the bands by SED
+    bg_cutoff: float
+        The minimum value in `detect` to include in detection.
+    """
+    B = observation.B
+    images = observation.get_scene(scene)
+    weights = np.array([sed[b]/bg_rms[b]**2 for b in range(B)])
+    jacobian = np.array([sed[b]**2/bg_rms[b]**2 for b in range(B)]).sum()
+    detect = np.einsum('i,i...', weights, images) / jacobian
+
+    # thresh is multiple above the rms of detect (weighted variance across bands)
+    bg_cutoff = thresh * np.sqrt((weights**2 * bg_rms**2).sum()) / jacobian
+    return detect, bg_cutoff
+
 
 def init_extended_source(sky_coord, scene, observations, bg_rms, obs_idx=0,
                          thresh=1., symmetric=True, monotonic=True):
@@ -85,7 +118,7 @@ def init_extended_source(sky_coord, scene, observations, bg_rms, obs_idx=0,
         observations = [observations]
     # determine initial SED from peak position
     sed = get_pixel_sed(sky_coord, scene, observations)  # amplitude is in sed
-    morph, bg_cutoff = observation.build_detection_coadd(sed, bg_rms, observations[obs_idx], scene, thresh)
+    morph, bg_cutoff = build_detection_coadd(sed, bg_rms, observations[obs_idx], scene, thresh)
     center = scene.get_pixel(sky_coord)
 
     # Apply the necessary constraints
